requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
oauth2client==4.1.3  # Для работы с GSC API
pandas==2.1.4
numpy>=1.21.0
pytest>=7.0.0
pytest-mock>=3.10.0
python-dotenv==1.0.0
supabase-py>=0.0.2
scikit-learn>=1.0.0  # для ML-моделей
psycopg2-binary==2.9.9
scipy==1.11.4
python-telegram-bot==20.7
aioschedule==0.5.2
PyYAML==6.0.1
httpx>=0.25.2
google-analytics-data==0.17.1
matplotlib>=3.8.0  # для визуализации отчетов
